Always busy-wait in emscripten_futex_wait on the main runtime thread
Even when the main runtime thread supports `Atomics.wait` we don't want
to use it since it can lead to deadlocks is we don't process incoming
events from other threads while we wait.

As an optimization here we could probably make it so that the main
thread can be woken up purely for the purposes of running a proxied
function and then goes back to sleep on the original mutex, but that
seems like a much larger change.

Fixes: #20059